Split into dedicated test classes & increase coverage.
