+++ source/runCoverage.sh
@@ -1,10 +1,13 @@
 #!/bin/sh
 
+rm -r ./**/TestResults
+rm -r ./coveragereport
+
 dotnet test --collect:"XPlat Code Coverage"
 
-reportgenerator -reports:'./**/coverage.*.xml' -targetdir:"coveragereport" -reporttypes:Html >/dev/null 2>&1
+reportgenerator -reports:'./**/coverage.*.xml' -targetdir:"coveragereport" -reporttypes:Html -verbosity:Error
 
 URL="file:coveragereport/index.html"
 [[ -x $BROWSER ]] && exec "$BROWSER" "$URL"
-path=$(which xdg-open || which gnome-open) && exec "$path" "$URL" >/dev/null 2>&1
-echo "Can't find browser"
+path=$(which xdg-open || which gnome-open) && exec "$path" "$URL" &
+exit 0
